Move ConstraintExtractor to plugin-toolkit
Rename ConstraintExtractor to UtcConstraintExtractor
